% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{as.rcat_series.data.frame}
\alias{as.rcat_series.data.frame}
\title{as.rcat_series.data.frame: convert  data.frame to rcat_series}
\usage{
\method{as.rcat_series}{data.frame}(x, start_time, interval = c(1440, 60, 30,
  10, 1), data_type_vals, ...)
}
\arguments{
\item{x}{\code{\link{data.frame}} 클래스의 변수}
}
\value{
\code{\link{POSIXt}} 형태의 시간 변수
}
\description{
\code{\link{data.frame}} 클래스에 부가적인 사항들을 덧붙여
\code{\link{rcat_series}} 클래스로 변환한다.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

