% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{as.rcat_series.xts}
\alias{as.rcat_series.xts}
\title{as.rcat_series.xts: get rcat_input StartTime}
\usage{
\method{as.rcat_series}{xts}(x, ...)
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스의 변수}
}
\value{
\code{\link{POSIXt}} 형태의 시간 변수
}
\description{
\code{\link{rcat_input}} 클래스의 StartTime 속성을 가져온다.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

