% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{as.xts.rcat_series}
\alias{as.xts.rcat_series}
\title{as.xts.rcat_series}
\usage{
as.xts.rcat_series(x, ...)
}
\arguments{
\item{x}{rcat 모형에서 사용하는 "rcat_series" 클래스의 변수}
}
\value{
변환된 \code{\link{xts}} 변수
}
\description{
"rcat_series" 클래스 자료를 xts로 변환하는 함수.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

