% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCATInput.R
\encoding{UTF-8}
\name{checkCATInput}
\alias{checkCATInput}
\title{checkCATInput}
\usage{
checkCATInput(x)
}
\arguments{
\item{x}{체크하려는 \code{\link{rcat_input}} 클래스의 변수나
\code{\link{readCATInput}} 함수 또는 \code{\link{readCATInput2}} 함수를
이용해 읽을 수 있는 파일의 경로}
}
\value{
체크 결과에 대한 \code{\link{character}} 변수, \code{\link{invisible}}
}
\description{
\code{\link{rcat_input}} 클래스와 연결된 파일 등의 무결성을 확인하는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

