% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2CATBinSeriesFile.R
\encoding{UTF-8}
\name{csv2CATBinSeriesFile}
\alias{csv2CATBinSeriesFile}
\title{csv2CATBinSeriesFile}
\usage{
csv2CATBinSeriesFile(csv, coltype, file)
}
\arguments{
\item{csv}{csv file name to read.}

\item{coltype}{\code{\link{seriesItemType}}}

\item{file}{file name converted to binary}
}
\description{
csv2CATBinSeriesFile
}
\examples{
csv2CATBinSeriesFile(csv = "climate3.csv",
  coltype = getSeriesItemTypeVals(date = 1,
                                  precip = 2,
                                  tempavg = 3,
                                  wind = 4,
                                  rhumi = 5,
                                  solar = 6),
  file = "climate3.dat")
}
\seealso{
\code{\link{seriesItemType}}
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
