% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_series.R
\encoding{UTF-8}
\name{getSeriesItemTypeVals}
\alias{getSeriesItemTypeVals}
\title{getSeriesItemTypeVals: get seriesItem type numbers.}
\usage{
getSeriesItemTypeVals(precip, eva, rhumi, solar, tempavg, wind, user, tempmin,
  tempmax, evacalc, date, observe)
}
\arguments{
\item{precip}{Column number of precipitation data}

\item{eva}{Column number of evapotranspiration data}

\item{rhumi}{Column number of relative humidity data}

\item{solar}{Column number of daylight hours data}

\item{tempavg}{Column number of average temperture data}

\item{wind}{Column number of windspeed data}

\item{user}{Column number of user data}

\item{tempmin}{Column number of minimum temperture data}

\item{tempmax}{Column number of maximum temperture data}

\item{evacalc}{Column number of calculated evapotranspiration data}

\item{date}{column number of date time data}

\item{observe}{관측값에 대한 컬럼 번호}
}
\value{
주어진 seriesItemType에 해당하는 정수 배열
}
\description{
series의 각 컬럼에 대한 seriesItem의 번호를 얻어오는 함수로서
\code{\link{as.rcat_series}}함수의 매개변수를 구하는데 사용됩니다.
}
\details{
각 변수별로 해당 컬럼의 번호를 입력하고 각 컬럼의 번호는 중복되면 안됩니다.

컬럼의 번호는 배열을 사용할 수 없으며 배열을 사용하면 해당 배열을 첫번째
숫자만 인식됩니다.
즉, 변수들은 테이블 상에 1회씩만 포함될 수 있음에 유의해야합니다.
}
\examples{
getSeriesItemTypeVals(precip = 1, tempavg = 2, rhumi = 3, wind = 4, solar = 5)
}
\seealso{
\code{\link{as.rcat_series}}
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
