% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYmD.R
\encoding{UTF-8}
\name{getYmDHMS}
\alias{getYmDHMS}
\alias{getYmDHM}
\title{getYmDHMS: get integer array about time}
\usage{
getYmDHMS(x)

getYmDHM(x)
}
\arguments{
\item{x}{\code{\link{POSIXt}}로 변환 될 수 있는 시간변수나 문자열}
}
\value{
연, 월, 일, 시, 분, 초와 같은 \code{\link{integer}} 배열
}
\description{
\code{\link{POSIXt}}로 변환 될 수 있는 시간변수나 문자열에서
연, 월, 일, 시, 분, 초와 같은 값을 얻어 정수 배열로 반환합니다.

getYmDHM: get integer array(years, mongths, days, hours, minutes) about time
}
\section{Functions}{
\itemize{
\item \code{getYmDHM}: get integer array about time
}}

\examples{
getYmDHMS(Sys.time())
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
