% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_input.R
\encoding{UTF-8}
\name{is.rcat_input}
\alias{is.rcat_input}
\title{is.rcat_input}
\usage{
is.rcat_input(x)
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스인지 확인 하고자하는 임의의 변수}
}
\value{
\code{\link{rcat_input}} 클래스 자료여부를 알려주는 \code{\link{logical}} 변수
}
\description{
\code{\link{rcat_input}} 클래스인지 아닌지 알려주는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

