% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{is.rcat_serieses}
\alias{is.rcat_serieses}
\title{is.rcat_serieses}
\usage{
is.rcat_serieses(x)
}
\arguments{
\item{x}{\code{\link{rcat_serieses}} 클래스인지 확인 하고자하는 임의의 변수}
}
\value{
\code{\link{rcat_serieses}} 클래스 자료여부를 알려주는 \code{\link{logical}} 변수
}
\description{
\code{\link{rcat_serieses}} 클래스인지 아닌지 알려주는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

