% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_diff.R
\name{list_by}
\alias{list_by}
\title{List all files by suffix (exclude all directories).}
\usage{
list_by(suffix, folder)
}
\arguments{
\item{suffix}{desired suffix}

\item{folder}{folder containing files}
}
\description{
List all files by suffix (exclude all directories).
}
