% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misscat.R
\name{missing}
\alias{missing}
\title{Dropping Missing Values}
\usage{
missing(df, k)
}
\arguments{
\item{df}{the input tibble}

\item{k}{threshold (ratio of missing values to drop the row or column)}
}
\value{
data frame
}
\description{
Drops rows or columns containing missing values if the number of the missing values exceeds a threshold
}
\examples{
missing(df, 0.2)
}
