% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rcat_input.R
\encoding{UTF-8}
\name{plot.rcat_input}
\alias{plot.rcat_input}
\title{plot.rcat_input: plot the sechmatic view of CAT model}
\usage{
\method{plot}{rcat_input}(x, ...)
}
\arguments{
\item{x}{rcat 모형에서 사용하는 "rcat_input" 클래스의 변수}

\item{...}{\code{\link{plot}} 함수 참조}
}
\description{
"rcat_input" 클래스 자료를 그래프 형태로 보여주는 함수.
}
\examples{
download.file("http://r.prj.kr/data/rcat/rcat_ex.zip", "rcat_ex.zip")
unzip("rcat_ex.zip")
model1 <- readCATInput("dist_101208.txt")
plot(model1)
model2 <- readCATInput2("dist_101208.txt")
plot(model2)
}
\author{
박희성 \email{hspark90@i-fam.net}
}

