% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{plot.rcat_series}
\alias{plot.rcat_series}
\title{plot.rcat_series: plot the rcat_series}
\usage{
\method{plot}{rcat_series}(x, ...)
}
\arguments{
\item{x}{rcat 모형에서 사용하는 "rcat_series" 클래스의 변수}
}
\description{
"rcat_series" 클래스 자료를 그래프 형태로 보여주는 함수.
}
\examples{
download.file("http://r.prj.kr/data/rcat/rcat_ex.zip", "rcat_ex.zip")
unzip("rcat_ex.zip")
iseri <- readBinCATSerieses("P_climate_1209.dat")
plot(iseri[[1]])
}
\author{
박희성 \email{hspark90@i-fam.net}
}

