% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\encoding{UTF-8}
\name{+.rcat_serieses}
\alias{+.rcat_serieses}
\title{+.rcat_serieses: connect two rcat_serieses}
\usage{
\method{+}{rcat_serieses}(e1, e2)
}
\arguments{
\item{e1}{\code{\link{rcat_serieses}} 클래스의 변수}

\item{e2}{\code{\link{rcat_serieses}} 클래스의 변수}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_series}} 클래스 자료를 더해 하나로 만들어 주는 연산자.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

