% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{print.rcat_input}
\alias{print.rcat_input}
\title{print.rcat_input: print rcat_input contents}
\usage{
\method{print}{rcat_input}(x, ...)
}
\arguments{
\item{x}{\code{\link{rcat_serieses}} 클래스의 변수}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_input}} 클래스 자료의 간단한 내용을 보여주는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

