% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\encoding{UTF-8}
\name{print.rcat_serieses}
\alias{print.rcat_serieses}
\title{print.rcat_serieses: print rcat_serieses contents}
\usage{
\method{print}{rcat_serieses}(x, ...)
}
\arguments{
\item{x}{\code{\link{rcat_serieses}} 클래스의 변수}

\item{...}{\code{\link{print}} 함수 참조}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_series}} 클래스 자료의 간단한 내용을 보여주는 함수
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
