% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_series.R
\docType{class}
\encoding{utf-8}
\name{rcat_series}
\alias{rcat_series}
\title{rcat_series class}
\description{
\code{\link{rcat_serieses}} class 를 구성하는 세부 class
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
\keyword{class}
