% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\encoding{UTF-8}
\name{rcat_serieses.rcat_series}
\alias{rcat_serieses.rcat_series}
\title{rcat_serieses.rcat_series: make rcat_serieses from rcat_series}
\usage{
\method{rcat_serieses}{rcat_series}(x, description, name, ...)
}
\arguments{
\item{x}{\code{\link{rcat_series}} 클래스의 변수}

\item{description}{Description of the \code{\link{rcat_serieses}} created.}

\item{name}{Name of the \code{\link{rcat_serieses}} created.}

\item{...}{\code{\link{rcat_serieses}} 함수 참조}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_series}} 로 부터 \code{\link{rcat_serieses}}를 생성하는 함수
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
