% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_series.R
\docType{class}
\encoding{utf-8}
\name{rcat_seriesItem}
\alias{rcat_seriesItem}
\title{rcat_seriesItem}
\description{
\code{\link{rcat_series}} class에 포함되는 Item class
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
\keyword{class}
