% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCATSeriesesBin.R
\encoding{UTF-8}
\name{readBinCATSerieses}
\alias{readBinCATSerieses}
\title{readBinCATSerieses: read the binary data of CAT model to R variable which is rcat_serieses class}
\usage{
readBinCATSerieses(filename)
}
\arguments{
\item{filename}{출력파일의 경로}
}
\value{
rcat 모형에서 사용하는 "rcat_serieses" 클래스의 변수
}
\description{
CAT 모형에서 사용하는 변수계열을 저장한 Binary 파일을 읽어오는 함수.
}
\examples{
download.file("http://r.prj.kr/data/rcat/rcat_ex.zip", "rcat_ex.zip")
unzip("rcat_ex.zip")

climate_data <- readBinCATSerieses("P_climate_1209.dat")
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
