% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_with_na.R
\name{rep_with_na}
\alias{rep_with_na}
\title{Replacing uninformative strings with NAs}
\usage{
rep_with_na(df)
}
\arguments{
\item{df}{the input data frame}
}
\value{
data frame
}
\description{
Replaces uninformative strings (eg. empty strings) with NAs
so they can be removed as missing values.
}
\examples{
rep_with_na(df)
}
