% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repwithna.R
\name{repwithna}
\alias{repwithna}
\title{Replacing uninformative strings with NAs}
\usage{
repwithna(df)
}
\arguments{
\item{df}{the input data frame}
}
\value{
data frame
}
\description{
Replaces uninformative strings (eg. empty strings) with NAs
so they can be removed as missing values.
}
\examples{
repwithna(df)
}
