% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_diff.R
\name{role_diff}
\alias{role_diff}
\title{Find Role Differences}
\usage{
role_diff(folder = ".", suffix = "xml")
}
\arguments{
\item{folder}{folder containing XML files (and where the output will go)}

\item{suffix}{Defaults to "XML"}
}
\value{
A XLSX file containing table with all the permission values
}
\description{
Exports table with permissions for all 2 or more XML files exported from the Define roles page in Moodle.
}
