% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{RunCAT}
\alias{RunCAT}
\title{RunCAT}
\usage{
RunCAT(infile = "", outfile = "", format = "")
}
\arguments{
\item{infile}{입력파일의 경로}

\item{outfile}{출력파일의 경로}

\item{format}{포맷}
}
\description{
RunCAT 모형 구동 함수.
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}

