% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCATInput.R
\encoding{UTF-8}
\name{saveCATInput}
\alias{saveCATInput}
\title{saveCATInput: save rcat_input object}
\usage{
saveCATInput(x, file)
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스의 변수}

\item{file}{저장할 파일의 이름(경로), \code{\link{character}}}
}
\value{
invisible NULL
}
\description{
\code{\link{rcat_input}} 클래스 자료를 저장하는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

