% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_series.R
\docType{data}
\encoding{UTF-8}
\name{seriesItemType}
\alias{seriesItemType}
\title{seriesItemType:  kinds of seriesItem}
\format{An object of class \code{list} of length 12.}
\usage{
seriesItemType
}
\description{
seriesItem의 종류를 구분하는 정수
}
\examples{
with(seriesItemType, c(precip, rhumi, tempavg, wind, solar))
}
\author{
박희성 \email{hspark90@i-fam.net}
}
\keyword{datasets}

