% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{data}
\encoding{UTF-8}
\name{seriesItemTypes}
\alias{seriesItemTypes}
\title{seriesItemTypes:  kinds of seriesItem}
\format{\preformatted{ chr [1:12] "precip" "eva" "rhumi" "solar" "tempavg" "wind" "user" "tempmin" ...
}}
\usage{
seriesItemTypes
}
\description{
seriesItem의 종류를 구분하는 문자열
}
\author{
박희성 \email{hspark90@i-fam.net}
}
\keyword{datasets}

