% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{StartTime<-.rcat_input}
\alias{StartTime<-.rcat_input}
\title{StartTime.rcat_input: set rcat_input StartTime}
\usage{
\method{StartTime}{rcat_input}(x) <- value
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스의 변수}

\item{value}{\code{\link{POSIXt}} 클래스의 변수}
}
\value{
\code{\link{ract_input}} 형태의 시간 변수
}
\description{
\code{\link{rcat_input}} 클래스의 StartTime 속성을 가져온다.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

