% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_input.R
\encoding{UTF-8}
\name{StartTime<-}
\alias{StartTime<-}
\title{set StartTime: set the start time of \code{\link{rcat_input}}}
\usage{
StartTime(x) <- value
}
\arguments{
\item{x}{설정대상 \code{\link{rcat_input}} 클래스의 변수}

\item{value}{\code{\link{POSIXt}} 클래스의 변수}
}
\description{
set the start time of \code{\link{rcat_input}}
}
