% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_input.R
\encoding{UTF-8}
\name{StartTime.rcat_input}
\alias{StartTime.rcat_input}
\title{StartTime.rcat_input: get rcat_input StartTime}
\usage{
\method{StartTime}{rcat_input}(x)
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스의 변수}
}
\value{
\code{\link{POSIXt}} 형태의 시간 변수
}
\description{
\code{\link{rcat_input}} 클래스의 StartTime 속성을 가져온다.
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
