% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_input.R
\encoding{UTF-8}
\name{StartTime}
\alias{StartTime}
\title{get StartTime: get the start time of \code{\link{rcat_input}}}
\usage{
StartTime(x)
}
\arguments{
\item{x}{\code{\link{rcat_input}}}
}
\description{
get the start time of \code{\link{rcat_input}}
}

