% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_input.R
\encoding{UTF-8}
\name{[.rcat_input}
\alias{[.rcat_input}
\title{[.rcat_input: Select from items of rcat_input}
\usage{
\method{[}{rcat_input}(x, i, ...)
}
\arguments{
\item{x}{\code{\link{rcat_input}} 클래스의 변수}

\item{i}{인덱스}
}
\value{
\code{\link{rcat_input}} 형태의 자료
}
\description{
\code{\link{rcat_input}} 클래스 자료에서 특정노드를 선택해 반환하는 연산자.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

