% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{[.rcat_serieses}
\alias{[.rcat_serieses}
\title{[.rcat_serieses: Select from items of rcat_serieses}
\usage{
\method{[}{rcat_serieses}(x, i, ...)
}
\arguments{
\item{x}{rcat_serieses 클래스의 변수}

\item{i}{인덱스}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
"rcat_series" 클래스 자료에서 특정노드를 선택해 반환하는 함수.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

