% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suscat.R
\name{suscat}
\alias{suscat}
\title{Detect suspected erroneous numeric data in user chosen columns}
\usage{
suscat(df, column = NULL, n = 5, num = "percent")
}
\arguments{
\item{df}{A dataframe or matrix like object}

\item{column}{vector like with names of columns to test}

\item{n}{default to 1, integer number or percentage of suspected values to identify}

\item{num}{either "number" if n should be read as a number of rows or 'percentage' if a percentage of total rows}
}
\value{
list of named vectors containing row numbers per column
}
\description{
Detect suspected erroneous numeric data in user chosen columns
}
\examples{
suscat(iris, col = c('Sepal.Width', 'Petal.Width'))
}
