% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcorr.R
\name{topcorr}
\alias{topcorr}
\title{Top correlated features}
\usage{
topcorr(df, k = 5)
}
\arguments{
\item{df}{the input tibble}

\item{k}{the number of top correlated features}
}
\value{
tibble
}
\description{
Create a tibble with the top k correlated pairs of features
}
\examples{
topcorr(df, 10)
}
