% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCATSeriesesBin.R
\encoding{UTF-8}
\name{writeBinCATSerieses}
\alias{writeBinCATSerieses}
\title{writeBinCATSerieses: write the CAT rcat_serieses class data to a binary file}
\usage{
writeBinCATSerieses(x, filename)
}
\arguments{
\item{x}{rcat 모형에서 사용하는 "rcat_serieses" 클래스의 변수}

\item{filename}{출력파일의 경로}
}
\value{
에러 메시지나 저장된 파일이름(보이지 않음)
}
\description{
CAT 모형에서 사용하는 변수계열을 Binary 파일로 저장하는 함수 .
}
\examples{
download.file("http://r.prj.kr/data/rcat/rcat_ex.zip", "rcat_ex.zip")
unzip("rcat_ex.zip")

climate_data <- readBinCATSerieses("P_climate_1209.dat")
writeBinCATSerieses(climate_data, "cldata.dat")
}
\author{
Heeseong Park \email{hspark90@i-fam.net}
}
