% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{as.rcat_series}
\alias{as.rcat_series}
\alias{as.rcat_series.data.frame}
\alias{as.rcat_series.matrix}
\alias{as.rcat_series.xts}
\title{as.rcat_series: convert  a object to rcat_series}
\usage{
as.rcat_series(x, start_time, interval = c(1440, 60, 30, 10, 1),
  data_type_vals, ...)

\method{as.rcat_series}{data.frame}(x, start_time, interval = c(1440, 60, 30,
  10, 1), data_type_vals, ...)

\method{as.rcat_series}{xts}(x, ...)

\method{as.rcat_series}{matrix}(x, ...)
}
\arguments{
\item{x}{변환하고자 하는 변수}

\item{start_time}{\code{\link{rcat_series}}의 시작시각}

\item{interval}{\code{\link{rcat_series}}의 시간간격}

\item{data_type_vals}{\code{\link{data.frame}}의 각 컬럼에 대한
\code{\link{seriesItemType}} 정수의 배열}
}
\value{
변환된 \code{\link{rcat_series}} 객체
}
\description{
주어진 객체에 부가적인 사항들을 덧붙여
\code{\link{rcat_series}} 클래스로 변환한다.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{xts}: 

\item \code{matrix}: 
}}
\author{
박희성 \email{hspark90@i-fam.net}
}
\seealso{
\code{\link{seriesItemType}}
}

