% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\encoding{UTF-8}
\name{c.rcat_serieses}
\alias{c.rcat_serieses}
\title{c.rcat_serieses: combine rcat_serieses}
\usage{
\method{c}{rcat_serieses}(...)
}
\arguments{
\item{...}{더하고자 하는 \code{\link{rcat_serieses}} 클래스 자료를 나열}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_serieses}} 클래스 자료를 더해 하나로 만드는 함수
}
\author{
박희성 \email{hspark90@i-fam.net}
}

