% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{getSeriesItemTypeVals}
\alias{getSeriesItemTypeVals}
\title{getSeriesItemTypeVals: convert seriesItem type string to seriesItem type number.}
\usage{
getSeriesItemTypeVals(type_strs)
}
\arguments{
\item{type_str}{seriesItemType을 구분하는 문자열 배열}
}
\value{
주어진 seriesItemType에 해당하는 정수
}
\description{
getSeriesItemTypeVals: convert seriesItem type string to seriesItem type number.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

