% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{getYmDHMS}
\alias{getYmDHM}
\alias{getYmDHMS}
\title{getYmDHMS: get integer array about time}
\usage{
getYmDHMS(x)

getYmDHM(x)
}
\arguments{
\item{x}{\code{\link{POSIXt}}로 변환 될 수 있는 시간변수나 문자열}
}
\value{
연, 월, 일, 시, 분, 초와 같은 \code{\link{integer}} 배열
}
\description{
\code{\link{POSIXt}}로 변환 될 수 있는 시간변수나 문자열에서
연, 월, 일, 시, 분, 초와 같은 값을 얻어 정수 배열로 반환합니다.

getYmDHM: get integer array about time
}
\section{Functions}{
\itemize{
\item \code{getYmDHM}: 
}}
\examples{
getYmDHMS(Sys.time())
}
\author{
박희성 \email{hspark90@i-fam.net}
}

