% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{package}
\encoding{utf-8}
\name{rcat-package}
\alias{rcat}
\alias{rcat-package}
\title{rcat-package: Run CAT Model on R Environment}
\description{
R 환경에서 CAT모형을 수행하기 위한 함수.
}
\details{
\tabular{ll}{
Package: \tab rcat\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-08-12\cr
License: \tab GPL-2\cr
}

CAT을 dll 형태로 컴파일하여 R 환경에서 스크립트 형태로 빠르게 실행할 수 있도록
작성된 함수입니다.
}
\author{
\href{http://www.i-fam.net/water/}{박희성} \email{hspark90@i-fam.net}
}
\keyword{package}

