% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\docType{class}
\encoding{utf-8}
\name{rcat_serieses}
\alias{rcat_serieses}
\title{rcat_serieses class}
\description{
CAT의 출력 파일을 읽어오거나 시계열 입력자료를 읽거나 저장 할 때 사용되는 클래스 입니다.
}
\author{
\href{http://www.i-fam.net/water/}{박희성} \email{hspark90@i-fam.net}
}
\keyword{class}

