% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat_serieses.R
\encoding{UTF-8}
\name{[.rcat_serieses}
\alias{[.rcat_serieses}
\title{[.rcat_serieses: Select from items of rcat_serieses}
\usage{
\method{[}{rcat_serieses}(x, i)
}
\arguments{
\item{x}{\code{\link{rcat_serieses}} 클래스의 변수}

\item{i}{인덱스}
}
\value{
\code{\link{rcat_serieses}} 형태의 자료
}
\description{
\code{\link{rcat_series}} 클래스 자료에서 특정노드를 선택해 반환하는 연산자.
}
\author{
박희성 \email{hspark90@i-fam.net}
}

