% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suscat.R
\name{suscat}
\alias{suscat}
\title{Detect suspected erroneous numeric data in user chosen columns}
\usage{
suscat(df, col = NULL, n = 0.01)
}
\arguments{
\item{df}{A dataframe or matrix like object}

\item{col}{list like with numbers of columns to test}

\item{n}{number or percentage of suspected values to identify}
}
\value{
list of named lists containing row numbers
}
\description{
Detect suspected erroneous numeric data in user chosen columns
}
\examples{
suscat(df, col = c(1,3,5), n = 10)
suscat(df, col = c(1,5,4,10), n = 0.02)
}
