% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suscat.R
\name{suscat}
\alias{suscat}
\title{Detect suspected erroneous numeric data in user chosen columns}
\usage{
suscat(df, column = NULL, n = 5, num = "percent")
}
\arguments{
\item{df}{A dataframe or matrix like object}

\item{n}{default to 1, integer number or percentage of suspected values to identify}

\item{num}{either "number" if n should be read as a number of rows or 'percentage' if a percentage of total rows}

\item{col}{vector like with names of columns to test}
}
\value{
list of named vectors containing row numbers per column
}
\description{
Detect suspected erroneous numeric data in user chosen columns
}
\examples{
suscat(df, col = c(1,3,5), n = 10)
suscat(df, col = c(1,5,4,10), n = 2, num = 'number')
}
