% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcorr.R
\name{topcorr}
\alias{topcorr}
\title{Top correlated features}
\usage{
topcorr(df, k = "all")
}
\arguments{
\item{df}{data.frame, the input data frame}

\item{k}{numeric or character, default at "all", the number of feature pairs to return}
}
\value{
data.frame, the sorted correlation data frame
}
\description{
Generates a data frame with the top k correlated pairs of features using pearson correlation
}
\examples{
topcorr(iris, 1)
}
