% Generated by roxygen2 (4.1.1): do not edit by hand
\encoding{UTF-8}
\name{writeBinCATSerieses}
\alias{writeBinCATSerieses}
\title{writeBinCATSerieses}
\usage{
writeBinCATSerieses(x, filename)
}
\arguments{
\item{x}{rcat 모형에서 사용하는 "rcat_serieses" 클래스의 변수}

\item{filename}{출력파일의 경로}
}
\value{
에러 메시지나 저장된 파일이름(보이지 않음)
}
\description{
CAT 모형에서 사용하는 변수계열을 Binary 파일로 저장하는 함수 .
}
\examples{
download.file("http://r.prj.kr/data/rcat/rcat_ex.zip", "rcat_ex.zip")
unzip("rcat_ex.zip")
}
\author{
박희성 \email{hspark90@i-fam.net}
}

